+++ CHANGELOG.md
@@ -6,6 +6,44 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [6.0.5] - 2024-06-11
+### Added
+- New Assistant UI - FAQs [#4089](https://github.com/rokwire/illinois-app/issues/4089).
+### Changed
+- Updated status texts for in-app event registration [#4086](https://github.com/rokwire/illinois-app/issues/4086).
+
+## [6.0.4] - 2024-06-10
+### Fixed
+- Fixed Explore Map markers build when all explores fall into the same position [#4098](https://github.com/rokwire/illinois-app/issues/4098).
+### Changed
+- Updated optional registration texts [#4043](https://github.com/rokwire/illinois-app/issues/4043).
+- Upgrade to Flutter 3.22.2.
+
+## [6.0.3] - 2024-06-07
+### Changed
+- Updated need to login message, reuse it across different app features [#4079](https://github.com/rokwire/illinois-app/issues/4079).
+- Cleaned up pronunciation control in Profile, prompt before delete [#4083](https://github.com/rokwire/illinois-app/issues/4083).
+### Added
+- New Assistant UI - in progress [#4089](https://github.com/rokwire/illinois-app/issues/4089).
+
+## [6.0.2] - 2024-06-04
+### Changed
+- Prompt to star event after creation [#3655](https://github.com/rokwire/illinois-app/issues/3655).
+- Cleanup settings directory content, rename and move panels to proper locations [#4029](https://github.com/rokwire/illinois-app/issues/4029).
+### Fixed
+- Fixed unhandled exception in refreshing HomeTwitterWidget.
+### Added
+- Added standard property "feature" in Analytics [#4029](https://github.com/rokwire/illinois-app/issues/4029).
+
+## [6.0.1] - 2024-05-21
+### Changed
+- Cleaned up "share" (ex. "qr") icon, align icon buttons in Event2HomePanel [#4041](https://github.com/rokwire/illinois-app/issues/4041).
+### Added
+- "My Canvas Courses" in the Academics panel [#4071](https://github.com/rokwire/illinois-app/issues/4071).
+
+## [6.0.0] - 2024-05-20
+- Increase version number to 6.0 [#4069](https://github.com/rokwire/illinois-app/issues/4069).
+
 ## [5.2.11] - 2024-05-16
 ### Changed
 - Updated messaged displayed on Event2DetailPanel Register/Unregister [#4042](https://github.com/rokwire/illinois-app/issues/4042).